extract container download into own module
preparation for fuse-containers
